requirements.txt
pandas
dhis2.py
azure-storage-blob
azure-identity
requests